requirements.txt
automata-lib
click